requirements.txt
aiocron
discord
python-dotenv
